Fix i18n initialization
