+++ i18n/es.po
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: i18next-conv\n"
-"POT-Creation-Date: 2020-05-27T06:18:40.167Z\n"
+"POT-Creation-Date: 2020-05-27T06:29:00.407Z\n"
 "PO-Revision-Date: 2018-10-25T09:02:35.143Z\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -49,11 +49,11 @@ msgstr "Visibilidad"
 msgid "Organisation Units visibility"
 msgstr "Visibilidad de unidades organizativas"
 
-msgid "User groups for Template Generation"
-msgstr "Grupos de usuario con acceso a la sección de generación"
+msgid "Access to Template Generation"
+msgstr "Acceso a la sección de generación de plantillas"
 
-msgid "User groups with access to Settings and Themes"
-msgstr "Grupos de usuario con acceso a la configuración y los temas"
+msgid "Access to Settings and Themes"
+msgstr "Acceso a la configuración y los temas"
 
 msgid "Data Set"
 msgstr "Data Set"
