+++ i18n/fr.po
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: i18next-conv\n"
-"POT-Creation-Date: 2020-05-27T06:18:40.167Z\n"
+"POT-Creation-Date: 2020-05-27T06:29:00.407Z\n"
 "PO-Revision-Date: 2019-10-10T10:41:21.864Z\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -47,10 +47,10 @@ msgstr ""
 msgid "Organisation Units visibility"
 msgstr ""
 
-msgid "User groups for Template Generation"
+msgid "Access to Template Generation"
 msgstr ""
 
-msgid "User groups with access to Settings and Themes"
+msgid "Access to Settings and Themes"
 msgstr ""
 
 msgid "Data Set"
