+++ src/index.js
@@ -1,7 +1,12 @@
 import { Provider } from "@dhis2/app-runtime";
+import i18n from "@dhis2/d2-i18n";
 import axios from "axios";
+import { init } from "d2";
+import { D2ApiDefault } from "d2-api";
+import _ from "lodash";
 import React from "react";
 import ReactDOM from "react-dom";
+import { AppContext } from "./webapp/contexts/api-context";
 import App from "./webapp/components/app/App";
 
 async function getBaseUrl() {
@@ -15,21 +20,49 @@ async function getBaseUrl() {
     }
 }
 
+const isLangRTL = code => {
+    const langs = ["ar", "fa", "ur"];
+    const prefixed = langs.map(c => `${c}-`);
+    return _(langs).includes(code) || prefixed.filter(c => code && code.startsWith(c)).length > 0;
+};
+
+const configI18n = ({ keyUiLocale: uiLocale }) => {
+    i18n.changeLanguage(uiLocale);
+    document.documentElement.setAttribute("dir", isLangRTL(uiLocale) ? "rtl" : "ltr");
+};
+
 async function main() {
-    const config = {
-        baseUrl: await getBaseUrl(),
-        apiVersion: "30",
-    };
+    const baseUrl = await getBaseUrl();
+
     try {
+        const d2 = await init({ baseUrl: baseUrl + "/api" });
+        const api = new D2ApiDefault({ baseUrl });
+        Object.assign(window, { bulkLoad: { d2, api } });
+
+        const userSettings = await api.get("/userSettings").getData();
+        configI18n(userSettings);
+
         ReactDOM.render(
-            <Provider config={config}>
-                <App />
+            <Provider config={{ baseUrl, apiVersion: "30" }}>
+                <AppContext.Provider value={{ d2, api }}>
+                    <App />
+                </AppContext.Provider>
             </Provider>,
             document.getElementById("root")
         );
     } catch (err) {
         console.error(err);
-        ReactDOM.render(<div>{err.toString()}</div>, document.getElementById("root"));
+        const message = err.toString().match("Unable to get schemas") ? (
+            <h3 style={{ margin: 20 }}>
+                <a rel="noopener noreferrer" target="_blank" href={baseUrl}>
+                    Login
+                </a>
+                {` ${baseUrl}`}
+            </h3>
+        ) : (
+                err.toString()
+            );
+        ReactDOM.render(<div>{message}</div>, document.getElementById("root"));
     }
 }
